Status code & response update

+++ src/services/reminders/reminders.js
@@ -1,4 +1,5 @@
 const db = require('../../app/firestore');
+const ResponseError = require('../../errors/response-error');
 const remindersCollection = db.collection('reminders');
 
 const {
@@ -7,17 +8,25 @@ const {
   updateUserById
 } = require('../users/user');
 
-async function getRemindersById(reminderId) {
+async function checkUserReminder(userId, reminderId) {
+  // Checks if the reminderId is in the user's reminders list
+  const userData = await getUserById(userId);
+  return (userData.reminders.includes(reminderId)) ? true : false;
+}
+
+async function getRemindersById(userId, reminderId) {
   try {
+    const isReminderValid = await checkUserReminder(userId, reminderId);
+    if (!isReminderValid) throw new ResponseError('Could not find the reminder for this user', 404);
+
+    // Get the reminder data
     const remindersDoc = await remindersCollection.doc(reminderId).get()
-    if (remindersDoc.empty) {
-      return null;
-    }
+    if (remindersDoc.empty) throw new ResponseError(404, 'Reminder not found');
     return remindersDoc.data();
   }
   catch (error) {
     console.error('Error getting reminders:', error);
-    throw new Error('Failed to get reminders by id from Firestore');
+    throw error
   }
 }
 
@@ -28,14 +37,14 @@ async function getRemindersByUserId(userId) {
     const remindersDocList = [];
 
     for (const reminderId of remindersIdList) {
-      const remindersDoc = await getRemindersById(reminderId);
-      if (remindersDoc) {
-        remindersDocList.push(remindersDoc);
+      const reminderDoc = await getRemindersById(userId, reminderId);
+      if (reminderDoc && reminderDoc.isActive) {
+        remindersDocList.push(reminderDoc);
       }
     }
-    if (remindersDocList.empty) {
-      return null;
-    }
+
+    if (remindersDocList.length === 0) return [];
+
     return remindersDocList;
   } catch (error) {
     console.error('Error getting reminders by user id:', error);
@@ -62,6 +71,10 @@ async function saveReminder(req) {
      */
     let reminderId = req.params.reminderId;
     if (reminderId) {
+      // Check if the reminderId is in the user's reminders list
+      const isReminderValid = await checkUserReminder(userId, reminderId);
+      if (!isReminderValid) throw new ResponseError('Could not find the reminder for this user', 404);
+
       await remindersCollection.doc(reminderId).update(newReminderData);
     } else {
       // "Creating" a new reminder
@@ -76,7 +89,7 @@ async function saveReminder(req) {
     };
   } catch (error) {
     console.error('Error creating reminder:', error);
-    throw new Error('Failed to create/edit reminder in Firestore');
+    throw error;
   }
 }
 
@@ -84,10 +97,15 @@ async function deleteReminder(req) {
   const reminderId = req.params.reminderId;
 
   try {
-    await remindersCollection.doc(reminderId).delete();
-
+    // Check if the reminderId is in the user's reminders list
     const userId = req.user.userId;
     const userData = await getUserById(userId);
+    if (!userData.reminders.includes(reminderId)) throw new ResponseError('Could not find the reminder for this user', 404);
+
+    // Delete the reminder
+    await remindersCollection.doc(reminderId).delete();
+
+    // Remove the reminderId from the user's reminders list
     let reminderIdList = userData.reminders;
     reminderIdList = reminderIdList.filter(id => id !== reminderId);
 
@@ -96,20 +114,25 @@ async function deleteReminder(req) {
       reminders: reminderIdList
     };
 
+    // Update the user's data with the new reminders list
     await updateUserById(userId, updatedUserData);
   } catch (error) {
     console.error('Error deleting reminder:', error);
-    throw new Error('Failed to delete reminder from Firestore')
+    throw error;
   }
 }
 
 async function finishReminder(req) {
-  const reminderId = req.params.reminderId;
   try {
+    // Check if the reminderId is in the user's reminders list
+    const reminderId = req.params.reminderId;
+    const userId = req.user.userId;
+    const isReminderValid = await checkUserReminder(userId, reminderId);
+    if (!isReminderValid) throw new ResponseError('Could not find the reminder for this user', 404);
+
     const reminderDocRef = remindersCollection.doc(reminderId);
     await reminderDocRef.update({ isActive: false });
 
-    
     const reminderDoc = await reminderDocRef.get();
     const reminderData = reminderDoc.data();
     return {
@@ -117,7 +140,7 @@ async function finishReminder(req) {
     }
   } catch (error) {
     console.error('Error finishing reminder:', error);
-    throw new Error('Failed to finish reminder in Firestore');
+    throw error;
   }
 }
